requirements.txt
groq==0.36.0
chromadb==1.3.5
pypdf==6.3.0
sentence-transformers==5.1.2
mcp[cli]==1.22.0
PyYAML==6.0.3
langgraph==1.0.3
langchain-core==1.0.7
langchain-groq==1.0.1
python-dotenv==1.2.1
pytest==9.0.1
lark==1.3.1
docling==2.63.0
easyocr==1.7.2
pydantic==2.12.4
requests==2.32.5
uvicorn==0.38.0
fastapi==0.121.3